+++ test-data/p1-preloaded_lists/personal.sql
@@ -1,0 +1,24 @@
+USE linkServicesGroup;
+
+INSERT INTO DomesticAnimal (domesticAnimalName) VALUES
+('Dog'), ('Cat'), ('Fish'), ('Bird'), ('Small Mammal'),
+('Reptile'), ('Farm Animal'), ('Insect'), ('Mouse / Rat');
+
+
+INSERT INTO Allergy (allergyName) VALUES
+('Respiratory (incl. Asthma)'), ('Skin'), ('Blood and Immune System'),
+('Circulatory System'), ('Digestive System'), ('Nervous System'),
+('Genitourinary System'), ('Drug-based');
+
+
+INSERT INTO FearPhobia (fearPhobiaName) VALUES
+('Spiders'), ('Snakes'), ('Heights'), ('Open Spaces'),
+('Dogs'), ('Thunder / Ligntning'), ('Closed Spaces'), ('Germs'),
+('Flying'), ('Holes'), ('Cancer'), ('Death'),
+('Public Speaking'), ('Being Alone'), ('Failure'), ('Birds'),
+('Chickens'), ('Crowds'), ('Intimacy'), ('Needles');
+
+
+INSERT INTO PersonalityTrait (traitName) VALUES
+('Outgoing'), ('Shy'), ('Nice'), ('Grouchy'), ('Serious'),
+('Playful'), ('Sloppy'), ('Neat'), ('Lazy'), ('Active');