+++ test-data/p1-preloaded_lists/technology.sql
@@ -1,0 +1,13 @@
+USE linkServicesGroup;
+
+INSERT INTO TechnologyForm (technologyName) VALUES
+('Smartphones'), ('Laptop Computers'),
+('Desktop Computers'), ('Tablets'),
+('Willing to learn');
+
+
+INSERT INTO GamingConsole (consoleName) VALUES
+('No Preference'), ('Willing to Learn'), ('Happy to Play'),
+('PC'), ('Smartphone'), ('Xbox'),
+('PlayStation'), ('Nintendo'), ('Retro Consoles'),
+('Arcade Machines');