+++ changelog.md
@@ -1,7 +1,26 @@
 # Changelog
 
-**./test-data/p1-preloaded_lists/geography.sql**
-* Added postcodes and suburbs test data.
-* Affects these tables:
-	* Postcode
-	* Location
+**./test-data/p1-preloaded_lists/certification.sql**
+* New file - Contains test data related to qualifications and experience.
+* Affected tables:
+	* CheckClearance
+	* Qualification
+	* ExperienceArea
+
+---
+
+**./test-data/p1-preloaded_lists/personal.sql**
+* New file - Contains test data related to personal attributes.
+* Affected tables:
+	* DomesticAnimal
+	* Allergy
+	* FearPhobia
+	* PersonalityTrait
+
+---
+
+**./test-data/p1-preloaded_lists/technology.sql**
+* New file - Contains test data related to technology.
+* Affected tables:
+	* TechnologyForm
+	* GamingConsole